chore: create @signalk/server-admin-ui package with repo
Move admin-ui to /packages/admin-ui and use it from npm
as a separate module instead of building it as part of
the server module. This makes it easier for this github
repo to be used as a "monorepo" and when developing/running
the server off git you don't need to build it.

Since admin-ui is its own module it should have its own
ignore files. This removes old rules from the root and
adds them to packages/admin-ui directory.

Docker build still retains the app-build intermediate
container build, which should be removed separately.