feat: Tcp provider without reconnect on idle (#581)
This commit reimplements tcp provider with reconnection logic
implemented with reconnect-core. This adds Fibonacci backoff with
maximum delay of 5 seconds and removes reconnection logic for
idle connections.

Previously tcp provider reconnected idle connections. This is not
needed, because as long as the tcp connection is alive we can assume
that the source is working. Repeated reconnections may be unwanted, as
when instruments are turned off and there is no data available.
OpenPlotter's misconfigured kplex left connections dangling in
CLOSE_WAIT and eventually caused kplex to run out of filedescriptors,
causing error logs to fill storage.