fix: some autodetect fixes (#737)
* fix: double output when noThrottle: true

We are piping the output two times, once in the constructor and
another time when handling the first line. This resulted in
every line being handled two times if noThrottle: true.

* fix: handle backpressure in autodetect

We were not handling backpressure properly: lots of input
would overwhelm the server, causing heap expansion. When write
returns false we need to defer calling done to when the output
emits 'drain'.

* refactor: simplify

As we are replacing _transform anyway we don't need to keep track
of the line being first or not, as the original _transform will
only get called once.

* refactor: use break in all cases