feature: more failure tolerant appstore
Appstore uses two registries. Now one can fail
and the store will still work with data from the
working one. The server waits for 5 seconds for the
slower registry to return and if not in time uses data
from the quicker one. The latest server version is
available only from npmjs - if it is not answering
the server check fails silently with nothing in the
ui.

Resolves #892.