feature: show plugin configure button (#1230)
This changes the plugin configuration logic so that the
enable/log/debug switches are not shown unless the plugin
has configuration data. Instead we show a Configure button
that opens the configuration form, that closes on first
submit. On first submit it also enables the plugin.

Fixes #1164 as the configuration form is closed on submit.