feature: clean up interfaces settings
Allow disabling only the interfaces that a normal user
might want to disable, and not all the interfaces. The
interfaces list structure is a relic with not much
relevance. You can still disable all the interfaces
by manually editing the settings file, but this makes
it so that you can not accidentally lock yourself
out of the admin UI.

Includes minor styling changes to get the switch labels
aligned a little more nicely.