fix: guard against plugin.schema() errors (#1365)
A single plugin that uses schema's function form whose schema() call
throws an error could prevent the server's plugin configuration ui
form page from loading.

This adds a guard against errors thrown in schema() call. The error
is logged and a fake schema containing a single field with a name
indicating that the schema could not be loaded. Problem with this
fake schema is that saving it may overwrite a perfectly good
plugin configuration, but it is a cheap way to show something is wrong.