feature: add per admin page Error Boundary
To prevent errors thrown in rendering from crashing the whole Admin UI
add an error boundary in conjunction with the login check. This way
bad data in localstorage or from the server is less likely to lock
the user out of the UI when the menu, restart and login are still
rendered.