feature: clear bad tokens in http requests
Previously bad tokens (whose verification fails for one reason or another)
were not cleared in http requests if read only access was on. After
 #1401 ws requests are closed with 401 status code if the token can not
be verified. This resultedin a state where the admin UI loaded but
websocket did not connect. This is not very useful, because read only
access over ws does not work and admin ui shows an annoying "Not
connected to the server" error message.

There is no use for bad tokens from cookies, so we can outright delete
the cookie whenever token verification fails. This handles for example
token expiration, falling back to read only access if it is available.

In browser use there are always http requests prior to opening the ws,
so bad token cookies are cleared before opening ws.