feature: add cr.signalk.io fly stuff
This allows we to refer to our images with a dns name that can
be changed if we want to move the images to some other registry.

So one would use this like

docker pull cr.signalk.io/signalk/signalk-server

This is already deployed and ghcr has one SK server image
tagged test so you can test this by appending :test to the
above.