Fix: multiplexed file detection & semicolon handling (#1661)
* fix: multiplexed file detection for plugin output

Logged output from plugins has the plugin id in the
discriminator field instead of the one letter data type
discriminator. If the first line of the multiplexed file
is from a plugin the file type detection does not work
and the file fails to play back silently.

Assert the minimum number of ; delimiters to be 3
instead of checking that charAt(15) is ;.

* fix: multiplexed file content with ;

If a multiplexed file line contains ; in the data, which is
perfectly possible for delta JSON, the part of the data
after ; would get thrown away.

Fix by joining the back with ; as delimiter.