fix: check ws max send buffer size over time
Fixes https://github.com/SignalK/signalk-server/issues/1718

Do not immediately hand ws up if the max send buffer size is
exceeded, but start monitoring for MAXSENDBUFFERCHECKTIME
milliseconds. If the buffer remains above the threshold all
the time terminate the connection. If the buffer has dropped
below the threshold during subsequent writes to the ws
connection the timer is reset.