doc: Fix broken link in plugin deprecation warning
The link in the deprecated `setProviderStatus`/`setProviderError` messages is broken, so this updates the message to explictly say what to use instead. It also switches to using the `deprecate` function from node's `util` module to do the deprecation, which support showing a stack trace.

Run `NODE_OPTIONS='--trace-deprecation' signalk-server` to see a full stack trace.